requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
